Move some useful bits for transcripts to annsec
